rpc: move Invoke to result package
It's just a data.
