+++ pkg/rpc/result/invoke.go
@@ -1,11 +1,12 @@
-package wrappers
+package result
 
 import (
 	"github.com/CityOfZion/neo-go/pkg/vm"
 )
 
-// InvokeResult is used as a wrapper to represent an invokation result.
-type InvokeResult struct {
+// Invoke represents code invocation result and is used by several RPC calls
+// that invoke functions, scripts and generic bytecode.
+type Invoke struct {
 	State       string `json:"state"`
 	GasConsumed string `json:"gas_consumed"`
 	Script      string `json:"script"`
